// src/libs/react-grid/reactgrid.d.ts
declare const isAlphaNumericKey: (keyCode: number) => boolean;
// ...
interface CellTemplate<TCell extends Cell = Cell> {
    /**
     * Validates and converts `uncertainCell` to compatible cell type
     *
     * @param {Uncertain<TCell>} uncertainCell Cell with all optional fields of its base (`TCell`)
     * @returns {Compatible<TCell>} Compatible cell of its base (`TCell`)
     */
    getCompatibleCell(uncertainCell: Uncertain<TCell>): Compatible<TCell>;
    /**
     * Returns `true` if the cell is focusable
     *
     * @param {Compatible<TCell>} cell Current cell as `Compatible` cell
     * @returns {boolean} `true` if cell should be focusable, by default returns `true`
     */
    isFocusable?(cell: Compatible<TCell>): boolean;
    /**
     * Updates cell based on new props. If not implemented, cell will be read-only
     *
     * @param {Compatible<TCell>} cell Current cell
     * @param {UncertainCompatible<TCell>} cellToMerge Incoming cell
     * @returns {Compatible<TCell>} Merged cell as `Compatible` cell
     */
    update?(cell: Compatible<TCell>, cellToMerge: UncertainCompatible<TCell>): Compatible<TCell>;
    /**
     * Handles keydown event on cell template and double click (opening cell in edit mode)
     * Default: cell => { cell, enableEditMode: false }
     *
     * @param {Compatible<TCell>} cell Incoming `Compatible` cell
     * @param {number} keyCode Represents the key pressed on the keyboard, or 1 for a pointer event (double click).
     * @param {boolean} ctrl Is `ctrl` pressed when event is called ()
     * @param {boolean} shift Is `shift` pressed when event is called
     * @param {boolean} alt Is `alt` pressed when event is called
     * @returns {{ cell: Compatible<TCell>; enableEditMode: boolean }} Cell data and edit mode either affected by the event or not
    */
    handleKeyDown?(cell: Compatible<TCell>, keyCode: number, ctrl: boolean, shift: boolean, alt: boolean): {
        cell: Compatible<TCell>;
        enableEditMode: boolean;
    };
    /**
     * Returns custom styles based on cell data applied to the cells `div` element
     * Default: _ => cell.style | {}
     *
     * @param {Compatible<TCell>} cell Incoming `Compatible` cell
     * @param {boolean} isInEditMode Flag is set to `true`, if cell is rendered in edit mode
     * @returns {CellStyle} Custom cell styling properties
     */
    getStyle?(cell: Compatible<TCell>, isInEditMode: boolean): CellStyle;
    /**
     *  Returns CSS classes based on cell data applied to the cells `div` element
     *
     * @param {Compatible<TCell>} cell Incoming `Compatible` cell
     * @param {boolean} isInEditMode Flag is set to `true`, if cell is rendered in edit mode
     * @returns {string} Cells CSS class names
     */
    getClassName?(cell: Compatible<TCell>, isInEditMode: boolean): string;
    /**
     * Renders the cell content
     *
     * @param {Compatible<TCell>} cell Incoming `Compatible` cell
     * @param {boolean} isInEditMode Flag is set to `true`, if cell is rendered in edit mode
     * @param {(cell: Compatible<TCell>, commit: boolean) => void} onCellChanged Callback used for commiting changes on a cell. For example: typing on html `input` element
     * @returns {React.ReactNode} Content of a cell
    */
    render(cell: Compatible<TCell>, isInEditMode: boolean, onCellChanged: (cell: Compatible<TCell>, commit: boolean) => void): React.ReactNode;
}
/**
 * `Id` is a common type to identify many ReactGrids objects
 *
 * @see https://reactgrid.com/docs/3.1/7-api/1-types/4-id/
 */
declare type Id = number | string;
// ...
interface BorderProps {
    /** Color of border - e.g. `#eee`/`red` */
    readonly color?: string;
    /** Style of border - e.g. `dotted`/`solid` */
    readonly style?: string;
    /** Width of border - e.g. `2px` */
    readonly width?: string;
}
/**
 * This interface styles single cell and prevents passing unwanted CSS properties that could break down grid rendering
 *
 * @see https://reactgrid.com/docs/3.1/7-api/0-interfaces/7-cell-style/
 */
interface CellStyle {
    /** CSS `color` property */
    readonly color?: string;
    /** CSS `background` property */
    readonly background?: string;
    /** CSS `overflow` property */
    readonly overflow?: string;
    /** CSS `padding-left` property */
    readonly paddingLeft?: string;
    /** Object that contains all cell's borders properties */
    readonly border?: {
        readonly left?: BorderProps;
        readonly top?: BorderProps;
        readonly right?: BorderProps;
        readonly bottom?: BorderProps;
    };
}
// ...
interface Cell {
    /** Name of cell type, must be unique */
    type: string;
    /** Marks cell as non editable (by default: `false`) */
    nonEditable?: boolean;
    /** `Id` of group to which this cell belongs to */
    groupId?: Id;
    /** Allowed style properties contained in `CellStyle` interface */
    style?: CellStyle;
    /** Additional CSS classes */
    className?: string;
}
/**
 * Cell type marker - every field of `TCell` is optional.
 * Cell of this type will have only one essential field provided by `Cell` interface - `type`.
 *
 * @see https://reactgrid.com/docs/3.1/7-api/1-types/2-uncertain-cell/
 */
declare type Uncertain<TCell extends Cell> = Partial<TCell> & Cell;
/**
 * Cell type marker - extended & exchangeable cell (compatible with different types)
 *
 * @see https://reactgrid.com/docs/3.1/7-api/1-types/1-compatible-cell/
 */
declare type Compatible<TCell extends Cell> = TCell & {
    /** Text value of a cell */
    text: string;
    /**  Numeric value of a cell, if there is no numeric value representation use `NaN` */
    value: number;
};
/**
 * `UncertainCompatible` is a cell type that is compatible with other cell types
 * that can be instances of various cell types (e.g. `DataCell` and `TimeCell`).
 *
 * @see https://reactgrid.com/docs/3.1/7-api/1-types/3-uncertain-compatible-cell/
 */
declare type UncertainCompatible<TCell extends Cell> = Uncertain<TCell> & {
    /** Text value of a cell */
    text: string;
    /** Numeric value of a cell, if there is no numeric value representation use `NaN` */
    value: number;
};
// ...
declare enum keyCodes {
    POINTER = 1,
    BACKSPACE = 8,
    TAB = 9,
    ENTER = 13,
    SHIFT = 16,
    CTRL = 17,
    ALT = 18,
    PAUSE = 19,
    CAPS_LOCK = 20,
    ESCAPE = 27,
    SPACE = 32,
    PAGE_UP = 33,
    PAGE_DOWN = 34,
    END = 35,
    HOME = 36,
    LEFT_ARROW = 37,
    UP_ARROW = 38,
    RIGHT_ARROW = 39,
    DOWN_ARROW = 40,
    INSERT = 45,
    DELETE = 46,
    KEY_0 = 48,
    KEY_1 = 49,
    KEY_2 = 50,
    KEY_3 = 51,
    KEY_4 = 52,
    KEY_5 = 53,
    KEY_6 = 54,
    KEY_7 = 55,
    KEY_8 = 56,
    KEY_9 = 57,
    KEY_A = 65,
    KEY_B = 66,
    KEY_C = 67,
    KEY_D = 68,
    KEY_E = 69,
    KEY_F = 70,
    KEY_G = 71,
    KEY_H = 72,
    KEY_I = 73,
    KEY_J = 74,
    KEY_K = 75,
    KEY_L = 76,
    KEY_M = 77,
    KEY_N = 78,
    KEY_O = 79,
    KEY_P = 80,
    KEY_Q = 81,
    KEY_R = 82,
    KEY_S = 83,
    KEY_T = 84,
    KEY_U = 85,
    KEY_V = 86,
    KEY_W = 87,
    KEY_X = 88,
    KEY_Y = 89,
    KEY_Z = 90,
    LEFT_META = 91,
    RIGHT_META = 92,
    SELECT = 93,
    NUMPAD_0 = 96,
    NUMPAD_1 = 97,
    NUMPAD_2 = 98,
    NUMPAD_3 = 99,
    NUMPAD_4 = 100,
    NUMPAD_5 = 101,
    NUMPAD_6 = 102,
    NUMPAD_7 = 103,
    NUMPAD_8 = 104,
    NUMPAD_9 = 105,
    MULTIPLY = 106,
    ADD = 107,
    SUBTRACT = 109,
    DECIMAL = 110,
    DIVIDE = 111,
    F1 = 112,
    F2 = 113,
    F3 = 114,
    F4 = 115,
    F5 = 116,
    F6 = 117,
    F7 = 118,
    F8 = 119,
    F9 = 120,
    F10 = 121,
    F11 = 122,
    F12 = 123,
    NUM_LOCK = 144,
    SCROLL_LOCK = 145,
    FIREFOX_DASH = 173,
    SEMICOLON = 186,
    EQUALS = 187,
    COMMA = 188,
    DASH = 189,
    PERIOD = 190,
    FORWARD_SLASH = 191,
    GRAVE_ACCENT = 192,
    OPEN_BRACKET = 219,
    BACK_SLASH = 220,
    CLOSE_BRACKET = 221,
    SINGLE_QUOTE = 222
}

/**
 * Gets property cell's value
 *
 * @param uncertainCell Cell to extract its property
 * @param propName Property name to extract
 * @param expectedType Expected `typeof` of extracted property from "uncertain" cell
 * @returns Value of property of given cell
 *
 * @see https://reactgrid.com/docs/3.1/7-api/2-functions/
 */
declare const getCellProperty: <TCell extends Cell, TKey extends keyof TCell>(uncertainCell: Uncertain<TCell>, propName: TKey, expectedType: 'string' | 'number' | 'boolean' | 'undefined' | 'function' | 'object' | 'symbol' | 'bigint') => NonNullable<Uncertain<TCell>[TKey]>;